+++ src/index.mjs
@@ -8,6 +8,8 @@ type ExpectChecks = {
   toNotEqual: (expected: mixed) => void,
   toThrow: (expectedError?: Error) => void,
   toNotThrow: () => void,
+  toMatch: (expected: string | RegExp) => void,
+  toNotMatch: (expected: string | RegExp) => void,
   toBeResolved: () => Promise<ExpectChecks>,
   toBeRejected: (expectedError?: Error) => Promise<void>,
 }
@@ -32,14 +34,14 @@ export function expect(given: mixed): ExpectChecks {
       const isEqual = equal(given, expected)
       if (!isEqual) {
         // $FlowFixMe
-        throw new BaumError(`${given} is not equal to ${expected.toString()}`)
+        throw new BaumError(`${given.toString()} is not equal to ${expected.toString()}`)
       }
     },
     toNotEqual(expected: mixed) {
       const isEqual = equal(given, expected)
       if (isEqual) {
         // $FlowFixMe
-        throw new BaumError(`${given} is equal to ${expected.toString()}`)
+        throw new BaumError(`${given.toString()} is equal to ${expected.toString()}`)
       }
     },
     toThrow(expectedError?: Error) {
@@ -61,9 +63,14 @@ export function expect(given: mixed): ExpectChecks {
       if (!receivedError) {
         throw new BaumError(`"${given.name}" do not throw an error!`)
       } else if (expectedError) {
-        throw new BaumError(`"${given.name}" throws error that do not equal to expected
+        if (
+          receivedError.name !== expectedError.name ||
+          receivedError.message !== expectedError.message
+        ) {
+          throw new BaumError(`"${given.name}" throws error that do not equal to expected
         Given name: ${receivedError.name} - expected name: ${expectedError.name}
           Given message: ${receivedError.message} - expected message: ${expectedError.message}`)
+        }
       }
     },
     toNotThrow() {
@@ -82,30 +89,36 @@ export function expect(given: mixed): ExpectChecks {
     },
     toMatch(expected: string | RegExp) {
       if (typeof given !== 'string') {
-        throw new TypeError(`Matched value must be of type "string", but given ${typeof given}`)
+        throw new TypeError(
+          `Matched value must be of type "string", but given ${typeof given}`
+        )
       }
 
-      const regexpedExpected = typeof expected === 'string'
-        ? new RegExp(expected)
-        : expected
+      const regexpedExpected =
+        typeof expected === 'string' ? new RegExp(expected) : expected
 
       const isMatched = regexpedExpected.test(given)
       if (!isMatched) {
-        throw new BaumError(`"${given}" does not match to ${regexpedExpected.toString()}`)
+        throw new BaumError(
+          `"${given}" does not match to ${regexpedExpected.toString()}`
+        )
       }
     },
     toNotMatch(expected: string | RegExp) {
       if (typeof given !== 'string') {
-        throw new TypeError(`Matched value must be of type "string", but given ${typeof given}`)
+        throw new TypeError(
+          `Matched value must be of type "string", but given ${typeof given}`
+        )
       }
 
-      const regexpedExpected = typeof expected === 'string'
-        ? new RegExp(expected)
-        : expected
+      const regexpedExpected =
+        typeof expected === 'string' ? new RegExp(expected) : expected
 
       const isMatched = regexpedExpected.test(given)
       if (isMatched) {
-        throw new BaumError(`"${given}" match to ${regexpedExpected.toString()}`)
+        throw new BaumError(
+          `"${given}" match to ${regexpedExpected.toString()}`
+        )
       }
     },
     async toBeResolved() {
@@ -162,7 +175,8 @@ function passed(title: string) {
 }
 
 function notPassed(title: string, error: BaumError) {
-  console.error(`   -  Test: "${title}" has errors!`)
-  console.error(error.toString())
-  console.warn('-- End of error --')
+  console.error(`   -  Test: "${title}" has errors!
+    : ${error.toString()}
+
+  ------ End of error ------`)
 }