+++ test/index.mjs
@@ -76,6 +76,18 @@ group('Self tests', () => {
     expect('sdf ').toNotMatch(/^[\s]+sdf$/)
   })
 
+  test('Test "toThrow()": function throws an error', () => {
+    expect(() => { throw new Error('Error') }).toThrow()
+  })
+
+  test('Test "toThrow()": function throws an error with message "Error"', () => {
+    expect(() => { throw new Error('Error') }).toThrow(new Error('Error'))
+  })
+
+  test('Test "toNotThrow()": function is not throw an error', () => {
+    expect(() => {}).toNotThrow()
+  })
+
   test('Test "toBeResolved()": Promise must toBeResolved properly', async () => {
     await expect(Promise.resolve(1)).toBeResolved()
   })